# docs/implementation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SOBA application API
  version: 2.0.0
  description: Normative implementation contract. All technical defaults are proposed
    in TD v2. Errors use Error. Unknown request fields are rejected. UUIDs use canonical
    lower-case form. See Backend-Spec.md for access and transaction rules.
servers:
- url: /
paths:
...
  /v1/auth/start:
    post:
      operationId: startLogin
      summary: startLogin
      description: Owner-scoped operation. See Backend-Spec.md for transaction and
        permission rules.
      security: []
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRedirect'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access or policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict or operation in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '410':
          description: Expired resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-requirements: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthStart'
  /v1/auth/mobile-exchange:
    post:
      operationId: exchangeMobileCode
      summary: exchangeMobileCode
      description: Owner-scoped operation. See Backend-Spec.md for transaction and
        permission rules.
      security: []
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tokens'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access or policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict or operation in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '410':
          description: Expired resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-requirements: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileExchange'
  /v1/auth/refresh:
    post:
      operationId: refreshMobileTokens
      summary: refreshMobileTokens
      description: Owner-scoped operation. See Backend-Spec.md for transaction and
        permission rules.
      security: []
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tokens'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access or policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict or operation in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '410':
          description: Expired resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-requirements: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Refresh'
...
  /v1/device-claims:
    post:
      operationId: createDeviceClaim
      summary: createDeviceClaim
      description: Owner-scoped operation. See Backend-Spec.md for transaction and
        permission rules.
      security:
      - sessionCookie: []
      - userBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id001
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          type: string
          minLength: 0
          maxLength: 128
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Claim'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access or policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict or operation in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '410':
          description: Expired resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-requirements:
      - U1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimCreate'
...
  /v1/device-claims/{id}/confirm:
    post:
      operationId: confirmDeviceClaim
      summary: confirmDeviceClaim
      description: Owner-scoped operation. See Backend-Spec.md for transaction and
        permission rules.
      security:
      - bootstrapBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id001
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceCredential'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access or policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict or operation in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '410':
          description: Expired resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-requirements:
      - U1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimConfirm'
...
  /v1/voice:
    get:
      operationId: upgradeVoiceSocket
      summary: upgradeVoiceSocket
      description: 'Unauthenticated handshake allowed only for app clients: first
        JSON event must redeem a ticket within 5 seconds. Device clients authenticate
        using deviceBearer before upgrade. No audio is accepted until session.ready.
        See Voice-Device-Spec.md.'
      security: []
      parameters: []
      responses:
        '101':
          description: Success
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access or policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict or operation in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '410':
          description: Expired resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-requirements: []
...
  /v1/invites:
    post:
      operationId: createInvite
      summary: createInvite
      description: Owner-scoped operation. See Backend-Spec.md for transaction and
        permission rules.
      security:
      - sessionCookie: []
      - userBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id001
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          type: string
          minLength: 0
          maxLength: 128
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invite'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access or policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict or operation in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '410':
          description: Expired resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Whole seconds. Included for rate-limited or request-in-progress
                outcomes.
              schema:
                type: integer
                minimum: 1
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-requirements:
      - U6
      - F1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteCreate'
...
components:
  securitySchemes:
    userBearer:
      type: http
      scheme: bearer
      description: SOBA opaque access token, 15 minute lifetime
    sessionCookie:
      type: apiKey
      in: cookie
      name: soba_session
    deviceBearer:
      type: http
      scheme: bearer
      description: Per-device operational credential
    bootstrapBearer:
      type: http
      scheme: bearer
      description: Per-device factory credential; claim confirmation only
    deletionReceipt:
      type: apiKey
      in: header
      name: X-Deletion-Receipt
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
          - unauthenticated
          - forbidden
          - not_found
          - invalid_request
          - version_conflict
          - idempotency_conflict
          - request_in_progress
          - rate_limited
          - dependency_unavailable
          - draft_expired
          - policy_blocked
          - claim_expired
          - device_owned
          - recipient_unavailable
          - invalid_state
          - unsupported_language
        message:
          type: string
          minLength: 0
          maxLength: 300
        request_id: *id001
        fields:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              field:
                type: string
                minLength: 0
                maxLength: 100
              code:
                type: string
                minLength: 0
                maxLength: 60
            required:
            - field
            - code
          maxItems: 20
      required:
      - code
      - message
      - request_id
      - fields
...
    AuthStart:
      type: object
      additionalProperties: false
      properties:
        client:
          type: string
          enum:
          - web
          - mobile
        return_path:
          type: string
          enum:
          - /app
          - /guardian
        mobile_challenge:
          type: string
          minLength: 43
          maxLength: 128
      required:
      - client
      - return_path
    AuthRedirect:
      type: object
      additionalProperties: false
      properties:
        authorization_url: *id017
        expires_at: *id013
      required:
      - authorization_url
      - expires_at
...
    Tokens:
      type: object
      additionalProperties: false
      properties:
        access_token: *id014
        refresh_token: *id014
        expires_in:
          type: integer
          minimum: 1
          maximum: 3600
      required:
      - access_token
      - refresh_token
      - expires_in
...